Update commercial license headers
Updated header.COMM to the files in tqtc-qtgamepad. Examples,
documentation files or tests are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4163.
Change-Id: Iaa9e71916cc8defcb21e09edcf6be9345fa732b5
Reviewed-by: Andy Nichols <andy.nichols@qt.io>